[#1] libs tests
