+++ go.work
@@ -1,4 +1,6 @@
-go 1.21.0
+go 1.22
+
+toolchain go1.22.0
 
 use (
 	.
