+++ internal/sqlx/repository.go
@@ -3,16 +3,23 @@ package sqlx
 import (
 	"context"
 	"fmt"
+	osqlx "github.com/kozmod/oniontx/sqlx"
 
-	"github.com/kozmod/oniontx-examples/internal/utils"
+	"github.com/kozmod/oniontx-examples/internal/entity"
+)
+
+type (
+	repoTransactor interface {
+		GetExecutor(ctx context.Context) osqlx.Executor
+	}
 )
 
 type TextRepository struct {
-	transactor    transactor
+	transactor    repoTransactor
 	errorExpected bool
 }
 
-func NewTextRepository(transactor transactor, errorExpected bool) *TextRepository {
+func NewTextRepository(transactor repoTransactor, errorExpected bool) *TextRepository {
 	return &TextRepository{
 		transactor:    transactor,
 		errorExpected: errorExpected,
@@ -21,7 +28,7 @@ func NewTextRepository(transactor transactor, errorExpected bool) *TextRepositor
 
 func (r *TextRepository) Insert(ctx context.Context, val string) error {
 	if r.errorExpected {
-		return utils.ErrExpected
+		return entity.ErrExpected
 	}
 	ex := r.transactor.GetExecutor(ctx)
 	_, err := ex.ExecContext(ctx, `INSERT INTO text (val) VALUES ($1)`, val)
