+++ internal/stdlib/usecase_test.go
@@ -2,16 +2,12 @@ package stdlib
 
 import (
 	"context"
-	"database/sql"
-	"fmt"
 	"testing"
 
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/stdlib"
-	oniontx "github.com/kozmod/oniontx/stdlib"
+	ostdlib "github.com/kozmod/oniontx/stdlib"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/kozmod/oniontx-examples/internal/utils"
+	"github.com/kozmod/oniontx-examples/internal/entity"
 )
 
 const (
@@ -31,7 +27,7 @@ func Test_UseCase(t *testing.T) {
 	t.Run("success_create", func(t *testing.T) {
 		var (
 			ctx         = context.Background()
-			transactor  = oniontx.NewTransactor(db)
+			transactor  = ostdlib.NewTransactor(db)
 			repositoryA = NewTextRepository(transactor, false)
 			repositoryB = NewTextRepository(transactor, false)
 			useCase     = NewUseCase(repositoryA, repositoryB, transactor)
@@ -57,7 +53,7 @@ func Test_UseCase(t *testing.T) {
 	t.Run("error_and_rollback", func(t *testing.T) {
 		var (
 			ctx         = context.Background()
-			transactor  = oniontx.NewTransactor(db)
+			transactor  = ostdlib.NewTransactor(db)
 			repositoryA = NewTextRepository(transactor, false)
 			repositoryB = NewTextRepository(transactor, true)
 			useCase     = NewUseCase(repositoryA, repositoryB, transactor)
@@ -65,7 +61,7 @@ func Test_UseCase(t *testing.T) {
 
 		err := useCase.CreateTextRecords(ctx, textRecord)
 		assert.Error(t, err)
-		assert.ErrorIs(t, err, utils.ErrExpected)
+		assert.ErrorIs(t, err, entity.ErrExpected)
 
 		{
 			records, err := GetTextRecords(db)
@@ -94,7 +90,7 @@ func Test_UseCases(t *testing.T) {
 		t.Run("success_create", func(t *testing.T) {
 			var (
 				ctx         = context.Background()
-				transactor  = oniontx.NewTransactor(db)
+				transactor  = ostdlib.NewTransactor(db)
 				repositoryA = NewTextRepository(transactor, false)
 				repositoryB = NewTextRepository(transactor, false)
 				useCases    = NewUseCases(
@@ -124,7 +120,7 @@ func Test_UseCases(t *testing.T) {
 		t.Run("error_and_rollback", func(t *testing.T) {
 			var (
 				ctx         = context.Background()
-				transactor  = oniontx.NewTransactor(db)
+				transactor  = ostdlib.NewTransactor(db)
 				repositoryA = NewTextRepository(transactor, false)
 				repositoryB = NewTextRepository(transactor, true)
 				useCases    = NewUseCases(
@@ -136,7 +132,7 @@ func Test_UseCases(t *testing.T) {
 
 			err := useCases.CreateTextRecords(ctx, textRecord)
 			assert.Error(t, err)
-			assert.ErrorIs(t, err, utils.ErrExpected)
+			assert.ErrorIs(t, err, entity.ErrExpected)
 
 			{
 				records, err := GetTextRecords(db)
@@ -150,44 +146,4 @@ func Test_UseCases(t *testing.T) {
 			})
 		})
 	})
-}
-
-func ConnectDB(t *testing.T) *sql.DB {
-	connConfig, err := pgx.ParseConfig(utils.ConnectionString)
-	assert.NoError(t, err)
-
-	connStr := stdlib.RegisterConnConfig(connConfig)
-	db, err := sql.Open("pgx", connStr)
-	assert.NoError(t, err)
-
-	err = db.Ping()
-	assert.NoError(t, err)
-
-	return db
-}
-
-func ClearDB(db *sql.DB) error {
-	_, err := db.Exec("TRUNCATE TABLE text;")
-	if err != nil {
-		return fmt.Errorf("clear DB: %w", err)
-	}
-	return nil
-}
-
-func GetTextRecords(db *sql.DB) ([]string, error) {
-	row, err := db.Query("SELECT val FROM text;")
-	if err != nil {
-		return nil, fmt.Errorf("get `text` records: %w", err)
-	}
-
-	var texts []string
-	for row.Next() {
-		var text string
-		err = row.Scan(&text)
-		if err != nil {
-			return nil, fmt.Errorf("scan `text` records: %w", err)
-		}
-		texts = append(texts, text)
-	}
-	return texts, nil
 }
